==> 2_model_growth_v_climate_stan.R <==
library(dplyr)
library(rstan)
library(foreach)
library(reshape2)
library(doParallel)
library(zoo) # for na.approx

load("growth_ctfsflagged_merged_detrended.RData")

growth <- tbl_df(growth)

growth <- filter(growth, ctfs_accept)
growth <- filter(growth, n_days > 200)
growth <- filter(growth, n_days < 550)

# Exclude NAK, CSN, and YAN - too little data at these sites
#growth <- filter(growth, !(sitecode %in% c("NAK", "CSN", "YAN")))

growth$SamplingPeriodID <- with(growth, factor(factor(sitecode):factor(SamplingPeriodNumber)))

###############################################################################
### TESTING ONLY
#growth <- filter(growth, sitecode %in% c("VB", "CAX"))
###############################################################################

n_chains <- 12
n_iter <- 1000

###############################################################################
#  Stan model

model_file <- "2_model_growth_v_climate_stan.stan" 



# Setup timeseries formatted dataframe (with NAs for covariates at time 1 since 
# first growth measurement can only be calculated from time 1 to time 2)
growth_wide <- data.frame(ID_tree=as.integer(factor(growth$SamplingUnitName)),
                          ID_plot=as.integer(factor(growth$plot_ID)),
                          ID_site=as.integer(factor(growth$sitecode)),
                          ID_period=as.integer(factor(growth$SamplingPeriodNumber)),
                          dbh_st=growth$diameter_start,
                          dbh_end=growth$diameter_end)
dbh_ts <- arrange(growth_wide, ID_period) %>%
    group_by(ID_tree) %>%
    do(data.frame(ID_tree=c(.$ID_tree[1], .$ID_tree),
                  ID_plot=c(.$ID_plot[1], .$ID_plot),
                  ID_site=c(.$ID_site[1], .$ID_site),
                  ID_period=c(1, .$ID_period + 1), # Start period IDs at 1
                  obs_num=seq(0, length(.$dbh_st)),
                  dbh=c(.$dbh_st[1], .$dbh_end))) %>%
    arrange(ID_tree, ID_period)

# Add latent growth inits
calc_latent_dbh <- function(dbh) {
    dbh_latent <- rep(NA, length(dbh))
    dbh_latent[1] <- dbh[1]
    for (i in 2:length(dbh)) {
        dbh_latent[i] <- ifelse(dbh[i] > dbh_latent[i - 1],
            dbh[i], dbh_latent[i - 1] + .01)
    }
    return(dbh_latent)
}
dbh_ts <- arrange(dbh_ts, ID_period) %>%
    group_by(ID_tree) %>%
    mutate(dbh_latent=calc_latent_dbh(dbh)) %>%
    arrange(ID_tree, ID_period)

# Setup wide format dbh and dbh_latent dataframes
dbh <- dcast(dbh_ts, ID_tree + ID_plot + ID_site ~ ID_period, value.var="dbh")
dbh_latent <- dcast(dbh_ts, ID_tree + ID_plot + ID_site ~ ID_period, value.var="dbh_latent")
ID_tree <- dbh$ID_tree
ID_plot <- dbh$ID_plot
ID_site <- dbh$ID_site
# Eliminate the ID columns
dbh <- dbh[!grepl('^ID_', names(dbh))]
dbh_latent <- dbh_latent[!grepl('^ID_', names(dbh_latent))]

dbh_missing <- is.na(dbh)

# Check how many trees have missing observations in the middle of their 
# observation periods (with each tree's observation period defined by the 
# periods with the first and last known diameter measurements).
calc_n_missings <- function(x) {
    first_obs <- min(which(!is.na(x)))
    last_obs <- max(which(!is.na(x)))
    sum(is.na(x[first_obs:last_obs]))
}
n_missing_periods <- t(apply(dbh, 1, calc_n_missings))
table(n_missing_periods)

# TEMPORARY: interpolate missing values in dbh (TODO: model NAs in Stan)
interp_dbh_obs <- function(x) {
    first_obs <- min(which(!is.na(x)))
    last_obs <- max(which(!is.na(x)))
    x[first_obs:last_obs] <- na.approx(as.numeric(x[first_obs:last_obs]))
    return(x)
}
dbh <- t(apply(dbh, 1, interp_dbh_obs))

# Interpolate missing values in dbh_latent
dbh_latent <- t(apply(dbh_latent, 1, interp_dbh_obs))

# Stan doesn't support NAs - but these rows will be ignored in Stan regardless 
# of their values since the indexing will be determined by "first_obs_period" and 
# "last_obs_period". So set them to 10 so that they will fit the constraints applied 
# to the matrix and not cause Stan to throw an error when they are read in.
dbh[is.na(dbh)] <- 10
dbh_latent[is.na(dbh_latent)] <- 10

# Setup data
n_tree <- length(unique(dbh_ts$ID_tree))
n_site <- length(unique(dbh_ts$ID_site))
n_plot <- length(unique(dbh_ts$ID_plot))

obs_per_tree <- group_by(dbh_ts, ID_tree) %>%
    summarize(first_obs_period=min(ID_period),
              last_obs_period=max(ID_period))

stan_data <- list(n_tree=n_tree,
                  n_plot=n_plot,
                  n_site=n_site,
                  first_obs_period=obs_per_tree$first_obs_period,
                  last_obs_period=obs_per_tree$last_obs_period,
                  max_obs_per_tree=ncol(dbh),
                  plot_ID=as.integer(factor(ID_plot)),
                  site_ID=as.integer(factor(ID_site)),
                  log_dbh=as.matrix(log(dbh)))
save(stan_data, file="stan_data.RData")

#var(log(dbh_ts$dbh))
# Setup inits
stan_init <- list(list(log_dbh_latent=as.matrix(log(dbh_latent)),
                       beta_0=-1.7,
                       beta_1=.008,
                       sigma_obs=.02, 
                       sigma_proc=.02, 
                       sigma_ijk=.06,
                       sigma_jk=.12,
                       sigma_k=.24,
                       b_ijk_std=rep(.5, n_tree),
                       b_jk_std=rep(.3, n_plot),
                       b_k_std=rep(.2, n_site)))
save(stan_init, file="stan_init.RData")

seed <- 1638
fit <- stan(model_file, data=stan_data, iter=n_iter, chains=1,
            init=rep(stan_init, 1), refresh=1, seed=seed)

# Fit initial model, on a single CPU. Run only one iteration.
stan_fit_initial <- stan(model_file, data=stan_data, iter=0, chains=1,
                         init=stan_init, chain_id=1)
print("finished running initial stan model")
save(stan_fit_initial, file="stan_fit_initial.RData")

# # Function to extract the last parameter estimates from a chain in a stanfit 
# # object to use as initialization values for a new chain.
# get_inits <- function(stan_fit, stan_init) {
#     pars <- names(stan_init[[1]])
#     init_param_ests <- extract(stan_fit_initial, permuted=TRUE, par=pars)
#     # Convert arrays into numeric
#     last_iter_num <- nrow(init_param_ests[[1]])
#     for (n in 1:length(init_param_ests)) {
#         # Select the last estimate if the chain has more than one iteration:
#         if (length(dim(init_param_ests[[n]])) > 1) {
#             init_param_ests[[n]] <- as.numeric(init_param_ests[[n]][last_iter_num, ])
#         } else {
#             init_param_ests[[n]] <- as.numeric(init_param_ests[[n]])
#         }
#     }
#     # Stan wants inits to be a list of lists:
#     init_param_ests <- list(init_param_ests)
#     return(init_param_ests)
# }
#
# # Extract the final values from this chain and use them to initialize the next
# # chains.
# stan_init_revised <- get_inits(stan_fit_initial, stan_init)
# save(stan_init_revised, file="stan_init_revised.RData")

# Fit n_chains chains in parallel. Reuse same seed so that the chain_ids can be 
# used by stan to properly seed each chain differently.
cl <- makeCluster(n_chains)
registerDoParallel(cl)
sflist <- foreach(n=1:n_chains, .packages=c("rstan")) %dopar% {
    # Add 1 to n in order to ensure chain_id 1 is not reused
    stan(fit=stan_fit_initial, data=stan_data, seed=seed, chains=1,
         iter=n_iter, chain_id=n+1, refresh=-1, init=stan_init)
    # stan(fit=stan_fit_initial, data=stan_data, seed=seed, chains=1,
    #      iter=n_iter, chain_id=n+1, refresh=-1, init=stan_init_revised)
}
print("finished running stan models on cluster")
stopCluster(cl)

stan_fit_allchains <- sflist2stanfit(sflist)
save(stan_fit_allchains, file="stan_fit_allchains.RData")

# model_params <- c("log_dbh_latent",
#                   "beta_0",
#                   "beta_1",
#                   "sigma_obs",
#                   "sigma_proc",
#                   "sigma_ijk",
#                   "sigma_jk",
#                   "sigma_k",
#                   "b_ijk",
#                   "b_jk",
#                   "b_k")
